+++ src/routes/analytics/categories/+page.svelte
@@ -1,11 +1,22 @@
 <script lang="ts">
+  import { onMount } from 'svelte';
+
   import { translate } from '$lib/translate';
   import Layout from '$lib/ui/layout/Layout.svelte';
 
   import AnalyticsBackButton from '../AnalyticsBackButton.svelte';
+  import AnalyticsBalancesVisibilityButton from '../AnalyticsBalancesVisibilityButton.svelte';
+  import { resetAnalyticsBalancesVisibility } from '../store';
   import CategoriesAnalytics from './CategoriesAnalytics.svelte';
+
+  onMount(resetAnalyticsBalancesVisibility);
 </script>
 
-<Layout title={$translate('categories.title')} leftSlot={AnalyticsBackButton}>
+<Layout
+  title={$translate('categories.title')}
+  leftSlot={AnalyticsBackButton}
+  rightSlot={AnalyticsBalancesVisibilityButton}
+  rightSlotProps={{ triState: true }}
+>
   <CategoriesAnalytics />
 </Layout>